backend/requirements.txt
letta-client
fastapi
uvicorn
python-dotenv
python-multipart
